requirements.txt
Django==1.8.4
django-appconf==1.0.1
django-clever-selects==0.6.9
django-crispy-forms==1.5.2
django-google-charts==0.1.1
django-haystack==2.4.0
django-localflavor==1.1
django-localflavor-us==1.1
django-model-utils==2.3.1
django-notifications-hq==0.7.0
django-qsstats-magic==0.7.2
django-registration-redux==1.2
django-smart-selects==1.1.1
elasticsearch==1.7.0
jsonfield==1.0.3
Pillow==3.0.0
python-dateutil==2.4.2
six==1.10.0
urllib3==1.12
wheel==0.24.0
